fix much stuff in the mem area
svn path=/trunk/kdelibs/kate/; revision=245167

+++ part/part/katebuffer.cpp
@@ -714,8 +714,8 @@ bool KateBuffer::needHighlight(KateBufBlock *buf, uint startLine, uint endLine)
     return false;
 
   TextLine::Ptr textLine;
-  QMemArray<uint> ctxNum, endCtx;
-  QMemArray<uchar> indentDepth;
+  QMemArray<short> ctxNum, endCtx;
+  QMemArray<unsigned short> indentDepth;
 
   uint last_line = buf->lines ();
   uint current_line = startLine - buf->startLine();
